Implement let polymorphism
- Also merge Solver and Constraints since they are mutually dependent
